third_party/blink/web_tests/wpt_internal/soft-navigation-heuristics/browser-initiated-popstate.tentative.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<title>Detect soft navigations done through the browser's back button.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/soft-navigation-heuristics/resources/soft-navigation-helper.js"></script>
</head>
<body>
  <main id=main>
    <div>
      <a id=link>Click me!</a>
    </div>
  </main>
  <script>
    // Push state twice, so that history.back() will trigger a popstate event,
    // when the first push state is restored.
    history.pushState({}, "", "foobar.html");
    history.pushState({}, "", "another_one.html");

    testSoftNavigation({
      testName: "A soft navigation that uses a *browser initiated* popstate"
        + " event is recognized by SoftNavigationHeuristics",
      interactionFunc: () => { testRunner.goToOffset(-1); },
      registerRouteChange: (cb) => addEventListener('popstate', cb),
    });
  </script>
</body>
</html>
