third_party/blink/web_tests/wpt_internal/soft-navigation-heuristics/browser-initiated-popstate.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] A soft navigation that uses a *browser initiated* popstate event is recognized by SoftNavigationHeuristics
  assert_equals: Soft navigation should not have triggered more LCP entries. expected 2 but got 1
Harness: the test ran to completion.
